backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
redis==5.0.1
pytest==7.4.3
black==23.10.1
isort==5.12.0
flake8==6.1.0
